🚚 Move CoilType class to base.constants.py and move the test function
